telegram_bot/pyproject.toml
[tool.poetry]
name = "bnt-telegram-bot"
version = "0.1.0"
description = "Telegram Bot for BNT Project"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.13.0"
aiohttp = "^3.10.5"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
redis = "^5.2.1"
openpyxl = "^3.1.5"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
mypy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 